+++ lib2/familia/settings.rb
@@ -3,11 +3,14 @@
 module Familia
 
   @delim = ':'
+  @prefix = nil
   @suffix = :object
+  @ttl = nil
+  @db = nil
 
   module Settings
 
-    attr_accessor :delim, :suffix
+    attr_accessor :delim, :suffix, :ttl, :db, :prefix
 
     # We define this do-nothing method because it reads better
     # than simply Familia.suffix in some contexts.
